+++ clang/lib/Basic/Targets/AVR.cpp
@@ -19,7 +19,7 @@
 using namespace clang;
 using namespace clang::targets;
 
-static constexpr int NumBuiltins = AVR::LastTSBuiltin - Builtin::FirstTSBuiltin;
+static constexpr int NumBuiltins = static_cast<int>(AVR::LastTSBuiltin) - static_cast<int>(Builtin::FirstTSBuiltin);
 
 static constexpr llvm::StringTable BuiltinStrings =
     CLANG_BUILTIN_STR_TABLE_START
