+++ clang/lib/Basic/Targets/DirectX.cpp
@@ -18,7 +18,7 @@ using namespace clang;
 using namespace clang::targets;
 
 static constexpr int NumBuiltins =
-    clang::DirectX::LastTSBuiltin - Builtin::FirstTSBuiltin;
+    static_cast<int>(clang::DirectX::LastTSBuiltin) - static_cast<int>(Builtin::FirstTSBuiltin);
 
 #define GET_BUILTIN_STR_TABLE
 #include "clang/Basic/BuiltinsDirectX.inc"
