+++ clang/lib/Basic/Targets/Mips.cpp
@@ -20,7 +20,7 @@ using namespace clang;
 using namespace clang::targets;
 
 static constexpr int NumBuiltins =
-    clang::Mips::LastTSBuiltin - Builtin::FirstTSBuiltin;
+    static_cast<int>(clang::Mips::LastTSBuiltin) - static_cast<int>(Builtin::FirstTSBuiltin);
 
 static constexpr llvm::StringTable BuiltinStrings =
     CLANG_BUILTIN_STR_TABLE_START
